# docs/person.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "person.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "PersonService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/persons": {
      "get": {
        "operationId": "PersonService_ListPerson",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListPersonResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "PersonService"
        ]
      },
      "post": {
        "operationId": "PersonService_CreatePerson",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreatePersonResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreatePersonRequest"
            }
          }
        ],
        "tags": [
          "PersonService"
        ]
      },
      "put": {
        "operationId": "PersonService_UpdatePerson",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUpdatePersonResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUpdatePersonRequest"
            }
          }
        ],
        "tags": [
          "PersonService"
        ]
      }
    },
    "/v1/persons/{id}": {
      "get": {
        "operationId": "PersonService_GetPerson",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetPersonResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "PersonService"
        ]
      },
      "delete": {
        "operationId": "PersonService_DeletePerson",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiDeletePersonResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "PersonService"
        ]
      }
    }
  },
  "definitions": {
    "apiCreatePersonRequest": {
      "type": "object",
      "properties": {
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        }
      },
      "title": "Create person"
    },
    "apiCreatePersonResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "apiDeletePersonResponse": {
      "type": "object"
    },
    "apiGetPersonResponse": {
      "type": "object",
      "properties": {
        "person": {
          "$ref": "#/definitions/apiPerson"
        }
      }
    },
    "apiListPersonResponse": {
      "type": "object",
      "properties": {
        "persons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiPerson"
          }
        }
      }
    },
    "apiPerson": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        }
      },
      "title": "Entity"
    },
    "apiUpdatePersonRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        }
      },
      "title": "Update person"
    },
    "apiUpdatePersonResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}